requirements.txt
tensorflow
opencv-python-headless
Pillow
streamlit
numpy
tqdm
gpt4all
gpt4all-j
streamlit-option-menu
